// apps/desktop/src/lib/generated/TarpackError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TarpackErrorKind } from "./TarpackErrorKind";

/**
 * `{ kind, message, entryId? }`.
 */
export type TarpackError = { kind: TarpackErrorKind, message: string, entryId?: string, };